Final commit before inital submission of project
Removed spaces from new projects (which broke it)
Added styling and fixed priority selection

Did not manage to work out how to repopulate the
project index AND have it display filter the new
category without reloading the page entirely.
Not ideal.

I also do not feel that I effectively used
factories or instances of classes. So it looks
like spaghetti and that's on me.